+++ dataflow/src/main/java/org/checkerframework/dataflow/expression/JavaExpression.java
@@ -878,7 +878,7 @@ public abstract class JavaExpression {
      */
     private static List<JavaExpression> argumentTreesToJavaExpressions(
             ExecutableElement method, List<? extends ExpressionTree> argTrees) {
-        if (isVarArgsInvocation(method, argTrees)) {
+        if (isVarargsInvocation(method, argTrees)) {
             List<JavaExpression> result = new ArrayList<>(method.getParameters().size());
             for (int i = 0; i < method.getParameters().size() - 1; i++) {
                 result.add(JavaExpression.fromTree(argTrees.get(i)));
@@ -908,7 +908,7 @@ public abstract class JavaExpression {
      * @return true if method is a varargs method and its varargs arguments are not passed in an
      *     array
      */
-    private static boolean isVarArgsInvocation(
+    private static boolean isVarargsInvocation(
             ExecutableElement method, List<? extends ExpressionTree> args) {
         if (!method.isVarArgs()) {
             return false;