+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -4616,9 +4616,9 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
                     overriddenParams.remove(0);
                 }
                 // Deal with varargs
-                if (overrider.isVarArgs() && !overridden.isVarArgs()) {
+                if (overrider.isVarargs() && !overridden.isVarargs()) {
                     overriderParams =
-                            AnnotatedTypes.expandVarArgsParametersFromTypes(
+                            AnnotatedTypes.expandVarargsParametersFromTypes(
                                     overrider, overriddenParams);
                 }
             }