+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeMirror.java
@@ -1299,7 +1299,7 @@ public abstract class AnnotatedTypeMirror implements DeepCopyable<AnnotatedTypeM
          *
          * @return true if this type represents a varargs method
          */
-        public boolean isVarArgs() {
+        public boolean isVarargs() {
             return this.element.isVarArgs();
         }
 
@@ -1334,7 +1334,7 @@ public abstract class AnnotatedTypeMirror implements DeepCopyable<AnnotatedTypeM
          *     excluding the receiver
          */
         /*package-private*/ void setParameterTypes(List<AnnotatedTypeMirror> params) {
-            if (paramTypesComputed && isVarArgs() && varargType == null) {
+            if (paramTypesComputed && isVarargs() && varargType == null) {
                 throw new BugInCF("Set vararg type before resetting parameter types");
             }
             paramTypes = params;
@@ -1415,7 +1415,7 @@ public abstract class AnnotatedTypeMirror implements DeepCopyable<AnnotatedTypeM
          * @param paramTypes the parameter types to determine the vararg type
          */
         private void computeVarargType(List<AnnotatedTypeMirror> paramTypes) {
-            if (!isVarArgs()) {
+            if (!isVarargs()) {
                 return;
             }
             varargType = (AnnotatedArrayType) paramTypes.get(paramTypes.size() - 1);