+++ framework/src/main/java/org/checkerframework/framework/type/GenericAnnotatedTypeFactory.java
@@ -2592,6 +2592,7 @@ public abstract class GenericAnnotatedTypeFactory<
     @FormatMethod
     private static void log(String format, Object... args) {
         if (debug) {
+            System.out.flush();
             SystemPlume.sleep(1); // logging can interleave with typechecker output
             System.out.printf(format, args);
         }