+++ framework/src/main/java/org/checkerframework/framework/util/typeinference8/types/InferenceFactory.java
@@ -1047,7 +1047,7 @@ public class InferenceFactory {
                 // TODO: the thrown types are not stored in the ExecutableElements, so the above
                 // method doesn't find any thrown types.  Below gets the types thrown type from the
                 // ExecutableType and just adds default annotations.  This is just a work around for
-                // this problem. We need to figure out how to get the type with the correct
+                // this problem.  We need to figure out how to get the type with the correct
                 // annotations.
                 List<AnnotatedTypeMirror> thrownTypesNew =
                         new ArrayList<>(thrownTypeMirrors.size());